+++ packages/zowe-explorer/__tests__/__unit__/dataset/actions.unit.test.ts
@@ -3858,3 +3858,26 @@ describe("Dataset Actions Unit Tests - Function allocateLike", () => {
         expect(errorHandlingSpy).toHaveBeenCalledWith(errorMessage, "test", "Unable to create data set.");
     });
 });
+
+describe("Dataset Actions Unit Tests - Function confirmJobSubmission", () => {
+    function createBlockMocks(): void {
+        mocked(vscode.window.showInputBox).mockImplementation((options) => {
+            options.validateInput("test");
+            return Promise.resolve("test");
+        });
+    }
+    it("Should use use local JCL doc name for confirmJobSubmission", async () => {
+        createGlobalMocks();
+        createBlockMocks();
+        jest.spyOn(vscode.workspace, "getConfiguration").mockImplementation(
+            () =>
+                ({
+                    get: () => sharedUtils.JOB_SUBMIT_DIALOG_OPTS[1],
+                } as any)
+        );
+        jest.spyOn(Gui, "warningMessage").mockResolvedValue({
+            title: "Submit",
+        });
+        await expect(dsActions.confirmJobSubmission("Profile\\test.jcl", true)).resolves.toEqual(true);
+    });
+});