+++ packages/zowe-explorer/__tests__/__unit__/job/actions.unit.test.ts
@@ -45,6 +45,7 @@ import { ZosJobsProvider } from "../../../src/job/ZosJobsProvider";
 import { ZoweLocalStorage } from "../../../src/utils/ZoweLocalStorage";
 import { LocalFileManagement } from "../../../src/utils/LocalFileManagement";
 import { ProfileManagement } from "../../../src/utils/ProfileManagement";
+import { TreeProviders } from "../../../src/shared/TreeProviders";
 
 jest.mock("../../../src/utils/LoggerUtils");
 
@@ -111,7 +112,8 @@ function createGlobalMocks() {
     Object.defineProperty(zowe.GetJobs, "getStatusForJob", { value: jest.fn(), configurable: true });
     Object.defineProperty(zowe.GetJobs, "getSpoolContentById", { value: jest.fn(), configurable: true });
     Object.defineProperty(vscode.workspace, "openTextDocument", { value: jest.fn(), configurable: true });
-    Object.defineProperty(vscode.window, "showTextDocument", { value: jest.fn(), configurable: true });
+    jest.spyOn(Gui, "showTextDocument");
+    jest.spyOn(vscode.window, "showTextDocument");
     Object.defineProperty(zowe, "ZosmfSession", { value: jest.fn(), configurable: true });
     Object.defineProperty(zowe.ZosmfSession, "createSessCfgFromArgs", { value: jest.fn(), configurable: true });
     Object.defineProperty(zowe, "DownloadJobs", { value: jest.fn(), configurable: true });
@@ -425,7 +427,7 @@ describe("Jobs Actions Unit Tests - Function downloadJcl", () => {
         await jobActions.downloadJcl(node);
         expect(mocked(zowe.GetJobs.getJclForJob)).toBeCalled();
         expect(mocked(vscode.workspace.openTextDocument)).toBeCalled();
-        expect(mocked(vscode.window.showTextDocument)).toBeCalled();
+        expect(mocked(Gui.showTextDocument)).toBeCalled();
     });
     it("Checking failed attempt to download Job JCL", async () => {
         createGlobalMocks();
@@ -567,6 +569,21 @@ describe("Jobs Actions Unit Tests - Function submitJcl", () => {
         const errorMsg = "No editor with a document that could be submitted as JCL is currently open.";
         expect(blockMocks.errorLogSpy).toBeCalledWith(errorMsg);
         expect(blockMocks.errorGuiMsgSpy).toBeCalledWith(errorMsg);
+    });
+
+    it("Checking cancel option scenario of local JCL submission confirmation dialog", async () => {
+        const blockMocks: any = createBlockMocks();
+        jest.spyOn(ZoweLogger, "trace").mockImplementation();
+        Object.defineProperty(vscode.window, "activeTextEditor", {
+            value: { document: { fileName: "test" } } as any,
+            configurable: true,
+        });
+        jest.spyOn(vscode.commands, "executeCommand").mockImplementation();
+        jest.spyOn(ZoweLogger, "debug").mockImplementation();
+        const confirmJobSubmissionSpy = jest.spyOn(dsActions, "confirmJobSubmission");
+        confirmJobSubmissionSpy.mockResolvedValue(false);
+        await expect(dsActions.submitJcl(blockMocks.testDatasetTree, {} as any)).resolves.toEqual(undefined);
+        confirmJobSubmissionSpy.mockRestore();
     });
 
     it("Checking failed attempt to submit of active text editor content as JCL without profile chosen from quickpick", async () => {
@@ -827,7 +844,7 @@ describe("Jobs Actions Unit Tests - Function submitMember", () => {
 });
 
 describe("Jobs Actions Unit Tests - Function getSpoolContent", () => {
-    function createBlockMocks() {
+    async function createBlockMocks() {
         const session = createISessionWithoutCredentials();
         const iJob = createIJobObject();
         const iJobFile = createIJobFile();
@@ -857,6 +874,11 @@ describe("Jobs Actions Unit Tests - Function getSpoolContent", () => {
             toJSON: jest.fn(),
         };
         bindJesApi(jesApi);
+        await TreeProviders.initializeProviders(null as any, {
+            ds: async (ctx) => null as any,
+            uss: async (ctx) => null as any,
+            job: async (ctx) => testJobTree,
+        });
 
         return {
             session,
@@ -874,23 +896,21 @@ describe("Jobs Actions Unit Tests - Function getSpoolContent", () => {
 
     it("should call showTextDocument with encoded uri", async () => {
         createGlobalMocks();
-        const blockMocks = createBlockMocks();
+        const blockMocks = await createBlockMocks();
         const session = "sessionName";
         const spoolFile = blockMocks.iJobFile;
-        const anyTimestamp = Date.now();
         mocked(SpoolProvider.encodeJobFile).mockReturnValueOnce(blockMocks.mockUri);
         mocked(Profiles.getInstance).mockReturnValue(blockMocks.profileInstance);
 
-        await jobActions.getSpoolContent(session, spoolFile, anyTimestamp);
+        await jobActions.getSpoolContent(session, { spool: spoolFile } as any);
 
-        expect(mocked(vscode.window.showTextDocument)).toBeCalledWith(blockMocks.mockUri, { preview: false });
+        expect(mocked(Gui.showTextDocument)).toBeCalledWith(blockMocks.mockUri, { preview: false });
     });
     it("should call showTextDocument with encoded uri with unverified profile", async () => {
         createGlobalMocks();
-        const blockMocks = createBlockMocks();
+        const blockMocks = await createBlockMocks();
         const session = "sessionName";
         const spoolFile = blockMocks.iJobFile;
-        const anyTimestamp = Date.now();
         Object.defineProperty(Profiles, "getInstance", {
             value: jest.fn(() => {
                 return {
@@ -905,13 +925,13 @@ describe("Jobs Actions Unit Tests - Function getSpoolContent", () => {
         mocked(SpoolProvider.encodeJobFile).mockReturnValueOnce(blockMocks.mockUri);
         mocked(Profiles.getInstance).mockReturnValue(blockMocks.profileInstance);
 
-        await jobActions.getSpoolContent(session, spoolFile, anyTimestamp);
+        await jobActions.getSpoolContent(session, { spool: spoolFile } as any);
 
-        expect(mocked(vscode.window.showTextDocument)).toBeCalledWith(blockMocks.mockUri, { preview: false });
+        expect(mocked(Gui.showTextDocument)).toBeCalledWith(blockMocks.mockUri, { preview: false });
     });
     it("should show error message for non existing profile", async () => {
         createGlobalMocks();
-        const blockMocks = createBlockMocks();
+        const blockMocks = await createBlockMocks();
         const session = "sessionName";
         const spoolFile = blockMocks.iJobFile;
         const anyTimestamp = Date.now();
@@ -920,14 +940,14 @@ describe("Jobs Actions Unit Tests - Function getSpoolContent", () => {
             throw new Error("Test");
         });
 
-        await jobActions.getSpoolContent(session, spoolFile, anyTimestamp);
+        await jobActions.getSpoolContent(session, { spool: spoolFile } as any);
 
         expect(mocked(vscode.window.showTextDocument)).not.toBeCalled();
         expect(mocked(Gui.errorMessage)).toBeCalledWith("Error: Test");
     });
     it("should show an error message in case document cannot be shown for some reason", async () => {
         createGlobalMocks();
-        const blockMocks = createBlockMocks();
+        const blockMocks = await createBlockMocks();
         const session = "sessionName";
         const spoolFile = blockMocks.iJobFile;
         const anyTimestamp = Date.now();
@@ -937,13 +957,13 @@ describe("Jobs Actions Unit Tests - Function getSpoolContent", () => {
             throw new Error("Test");
         });
 
-        await jobActions.getSpoolContent(session, spoolFile, anyTimestamp);
+        await jobActions.getSpoolContent(session, { spool: spoolFile } as any);
 
         expect(mocked(Gui.errorMessage)).toBeCalledWith("Error: Test");
     });
     it("should fetch the spool content successfully", async () => {
         createGlobalMocks();
-        const blockMocks = createBlockMocks();
+        const blockMocks = await createBlockMocks();
         const testNode = new Job(
             "undefined:test - testJob",
             vscode.TreeItemCollapsibleState.None,